+++ tests/spec/templates/schema.ts
@@ -1647,13 +1647,17 @@ export class HttpClient<SecurityDataType = unknown> {
 }
 
 /* HTTP CLIENT TEMPLATE */
+
 /**
  * @title GitHub
  * @version v3
+ * @termsOfService https://help.github.com/articles/github-terms-of-service/#b-api-terms
  * @baseUrl https://api.github.com
+ * @externalDocs https://developer.github.com/v3/
+ *
  * Powerful collaboration, code review, and code management for open source and private projects.
  */
-export class Api<SecurityDataType = any> extends HttpClient<SecurityDataType> {
+export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
   emojis = {
     /**
      * @description Lists all the emojis available to use on GitHub.